homework for lesson 7
