+++ myapp/src/App.js
@@ -1,13 +1,16 @@
 import "./App.css";
 import { Routers } from "./routes";
 import { Provider } from 'react-redux';
-import { store } from './store';
+import { persistor, store } from './store';
+import { PersistGate } from "redux-persist/es/integration/react"
 
 function App() {
 
   return (
     <Provider store={store}>
-      <Routers />
+      <PersistGate persistor={persistor} >
+        <Routers />
+      </PersistGate>
     </Provider>
   )
 }
