+++ myapp/src/store/messages/actions.js
@@ -18,3 +18,4 @@ export const deleteMessage = (chatId, id) => ({
         id
     },
 });
+// let timeout;
